requirements.txt
discord_py_interactions==5.12.0
python-dotenv==1.0.1